+++ arbolBB.h
@@ -1,6 +1,6 @@
 
 struct nodo{//Estructura de la conformación de los nodos
-    int dato;
+    long int dato;
     struct nodo *hijoIzq;
     struct nodo *hijoDer;
 };
@@ -11,6 +11,6 @@ struct arbolBB{//Nodo arbol, este será la raíz
 
 //Declaración de las operaciones del arbol 
 struct arbolBB *crearABB();
-struct nodo *crearNodo(int dato);
-void insertar(struct arbolBB *arbol, int dato);
-void guardarRecorridoInorden(struct arbolBB *arbol, int *A, int n);
+struct nodo *crearNodo(long int dato);
+void insertar(struct arbolBB *arbol, long int dato);
+void guardarRecorridoInorden(struct arbolBB *arbol, long int *A, long int n);
